+++ examples/snap-chapi-example/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @learncard/snap-chapi-example
 
+## 1.0.217
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.22
+    -   @learncard/helpers@1.1.11
+    -   @learncard/react@2.7.73
+    -   @learncard/meta-mask-snap@1.0.73
+
+## 1.0.216
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.21
+    -   @learncard/helpers@1.1.10
+    -   @learncard/react@2.7.72
+    -   @learncard/meta-mask-snap@1.0.72
+
 ## 1.0.215
 
 ### Patch Changes