+++ packages/plugins/didkey/CHANGELOG.md
@@ -1,5 +1,21 @@
 # learn-card-core
 
+## 1.0.33
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.22
+    -   @learncard/helpers@1.1.11
+
+## 1.0.32
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.21
+    -   @learncard/helpers@1.1.10
+
 ## 1.0.31
 
 ### Patch Changes