+++ packages/plugins/learn-cloud/CHANGELOG.md
@@ -1,5 +1,25 @@
 # learn-card-core
 
+## 2.1.17
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.22
+    -   @learncard/helpers@1.1.11
+    -   @learncard/learn-cloud-client@1.3.16
+    -   @learncard/didkit-plugin@1.5.8
+
+## 2.1.16
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.21
+    -   @learncard/helpers@1.1.10
+    -   @learncard/learn-cloud-client@1.3.15
+    -   @learncard/didkit-plugin@1.5.7
+
 ## 2.1.15
 
 ### Patch Changes