+++ packages/plugins/simple-signing-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # lca-api-plugin
 
+## 1.0.5
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/simple-signing-client@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/simple-signing-client@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes