+++ services/meta-mask-snap/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @learncard/meta-mask-snap
 
+## 1.0.73
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.22
+
+## 1.0.72
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.21
+
 ## 1.0.71
 
 ### Patch Changes